+++ src/main/java/uk/gov/crowncommercial/dts/scale/cat/controller/MessageController.java
@@ -8,7 +8,8 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import uk.gov.crowncommercial.dts.scale.cat.model.generated.Message;
+import uk.gov.crowncommercial.dts.scale.cat.model.generated.*;
+import uk.gov.crowncommercial.dts.scale.cat.model.jaggaer.MessageRequestInfo;
 import uk.gov.crowncommercial.dts.scale.cat.service.MessageService;
 
 /**
@@ -16,7 +17,7 @@ import uk.gov.crowncommercial.dts.scale.cat.service.MessageService;
  *
  */
 @RestController
-@RequestMapping(path = "/tenders/projects", produces = APPLICATION_JSON_VALUE)
+@RequestMapping(path = "/tenders/projects/{proc-id}/events/{event-id}/messages", produces = APPLICATION_JSON_VALUE)
 @RequiredArgsConstructor
 @Slf4j
 @Validated
@@ -24,7 +25,7 @@ public class MessageController extends AbstractRestController {
 
   private final MessageService messageService;
 
-  @PostMapping("{proc-id}/events/{event-id}/messages")
+  @PostMapping()
   public ResponseEntity<String> createAndRespondMessage(
       @PathVariable("proc-id") final Integer procId, @PathVariable("event-id") final String eventId,
       @Valid @RequestBody final Message messageRequest,
@@ -38,5 +39,45 @@ public class MessageController extends AbstractRestController {
     return ResponseEntity
         .ok(messageService.sendOrRespondMessage(principal, procId, eventId, messageRequest));
   }
+  @GetMapping()
+  public MessageSummary getMessages(
+          @PathVariable("proc-id") final Integer procId,
+          @PathVariable("event-id") final String eventId,
+          @RequestParam(name = "message-direction",required = false, defaultValue = "RECEIVED") MessageDirection messageDirection,
+          @RequestParam(name = "message-read",required = false, defaultValue = "ALL") MessageRead messageRead,
+          @RequestParam(name = "sort",required = false, defaultValue = "DATE") MessageSort messageSort,
+          @RequestParam(name = "sort-order",required = false , defaultValue = "ASCENDING") MessageSortOrder messageSortOrder,
+          @RequestParam(name = "page",required = false, defaultValue = "1") Integer page,
+          @RequestParam(name = "page-size",required = false, defaultValue = "20") Integer pageSize,
+          final JwtAuthenticationToken authentication) {
 
+    var principal = getPrincipalFromJwt(authentication);
+    log.info("getMessagesSummaries invoked on behalf of principal: {}", principal);
+
+    var  messageRequestInfo = MessageRequestInfo.builder()
+            .procId(procId)
+            .eventId(eventId)
+            .messageDirection(messageDirection)
+            .messageRead(messageRead)
+            .messageSort(messageSort)
+            .messageSortOrder(messageSortOrder)
+            .page(page)
+            .pageSize(pageSize)
+            .principal(principal)
+            .build();
+    return messageService.getMessagesSummary(messageRequestInfo);
+  }
+
+  @GetMapping("/{message-id}")
+  public Message getMessage(
+          @PathVariable("proc-id") final Integer procId,
+          @PathVariable("event-id") final String eventId,
+          @PathVariable("message-id") final String messageId,
+          final JwtAuthenticationToken authentication) {
+
+    var principal = getPrincipalFromJwt(authentication);
+    log.info("getMessagesSummaries invoked on behalf of principal: {}", principal);
+
+    return messageService.getMessageSummary(procId, eventId,messageId, principal);
+  }
 }
