+++ src/main/java/uk/gov/crowncommercial/dts/scale/cat/service/MessageService.java
@@ -1,8 +1,10 @@
 package uk.gov.crowncommercial.dts.scale.cat.service;
 
 import static org.springframework.util.CollectionUtils.isEmpty;
-import java.util.HashMap;
-import java.util.Map;
+
+import java.net.URI;
+import java.util.*;
+import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import org.apache.commons.collections.CollectionUtils;
 import org.springframework.boot.configurationprocessor.json.JSONObject;
@@ -15,11 +17,13 @@ import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import uk.gov.crowncommercial.dts.scale.cat.config.RPAAPIConfig;
+import uk.gov.crowncommercial.dts.scale.cat.exception.AuthorisationFailureException;
 import uk.gov.crowncommercial.dts.scale.cat.exception.JaggaerRPAException;
 import uk.gov.crowncommercial.dts.scale.cat.model.entity.OrganisationMapping;
 import uk.gov.crowncommercial.dts.scale.cat.model.entity.ProcurementEvent;
-import uk.gov.crowncommercial.dts.scale.cat.model.generated.Message;
-import uk.gov.crowncommercial.dts.scale.cat.model.generated.MessageNonOCDS;
+import uk.gov.crowncommercial.dts.scale.cat.model.generated.*;
+import uk.gov.crowncommercial.dts.scale.cat.model.jaggaer.MessageRequestInfo;
+import uk.gov.crowncommercial.dts.scale.cat.model.jaggaer.Receiver;
 import uk.gov.crowncommercial.dts.scale.cat.model.rpa.*;
 import uk.gov.crowncommercial.dts.scale.cat.repo.RetryableTendersDBDelegate;
 
@@ -35,6 +39,8 @@ public class MessageService {
 
   private static final String RESPOND_MESSAGE = "Respond";
 
+  public static final String JAGGAER_USER_NOT_FOUND = "Jaggaer user not found";
+
   private final WebClient rpaServiceWebClient;
 
   private final WebclientWrapper webclientWrapper;
@@ -49,7 +55,16 @@ public class MessageService {
 
   private final JaggaerService jaggaerService;
 
+  private final UserProfileService userProfileService;
+
   private final RetryableTendersDBDelegate retryableTendersDBDelegate;
+
+  private static final String LINK_URI = "/messages/";
+
+  private static final int RECORDS_PER_REQUEST = 100;
+
+  private final Comparator<uk.gov.crowncommercial.dts.scale.cat.model.jaggaer.Message> comparator =
+          Comparator.comparing(o -> o.getSender().getName());
 
   /**
    * Which sends outbound message to all suppliers and single supplier. And also responds supplier
@@ -156,7 +171,7 @@ public class MessageService {
   }
 
   /**
-   * @param processInputMap
+   * @param processInput
    * @return rpa status
    * @throws JsonProcessingException
    */
@@ -225,5 +240,180 @@ public class MessageService {
     var uriTemplate = rpaAPIConfig.getAuthenticationUrl();
     return webclientWrapper.postData(jaggerRPACredentials, String.class, rpaServiceWebClient,
         rpaAPIConfig.getTimeoutDuration(), uriTemplate);
+  }
+
+  /**
+   * Returns a list of message summary at the event level.
+   * @param messageRequestInfo
+   * @return
+   */
+  public MessageSummary getMessagesSummary(final MessageRequestInfo messageRequestInfo) {
+
+    var jaggaerUserId = userProfileService.resolveBuyerUserByEmail(messageRequestInfo.getPrincipal())
+            .orElseThrow(() -> new AuthorisationFailureException(JAGGAER_USER_NOT_FOUND)).getUserId();
+    var event = validationService.validateProjectAndEventIds(messageRequestInfo.getProcId(),
+            messageRequestInfo.getEventId());
+
+    Predicate<uk.gov.crowncommercial.dts.scale.cat.model.jaggaer.Message> directionPredicate = message ->
+            (MessageDirection.ALL.equals(messageRequestInfo.getMessageDirection())
+                    || message.getDirection().equals(messageRequestInfo.getMessageDirection().getValue()));
+    Predicate<Receiver> receiverPredicate = receiver ->  MessageRead.ALL.equals(messageRequestInfo.getMessageRead())
+            || ((MessageRead.READ.equals(messageRequestInfo.getMessageRead())
+            && receiver.getId().equals(jaggaerUserId)));
+
+    var messagesResponse = jaggaerService.getMessages(event.getExternalReferenceId(),
+            messageRequestInfo.getPage());
+    var allMessages = messagesResponse.getMessageList().getMessage();
+
+    /**
+     * Make first request to jagger
+     * if total records are more than  > 100 (Jaggaer returns max 100 order by date desc) and messageSort is TITLE/AUTHOR
+     * then make sub-sequent call to get total records, then messageSort it
+     * send only requested no of records
+     *
+     *
+     */
+    if (MessageSort.AUTHOR.equals(messageRequestInfo.getMessageSort())
+            || MessageSort.TITLE.equals(messageRequestInfo.getMessageSort())) {
+      var totalRecords = messagesResponse.getTotRecords();
+      var pages = Math.round(totalRecords / (double) RECORDS_PER_REQUEST);
+      for (int i = 2; i < pages; i++) {
+        var response = jaggaerService.getMessages(event.getExternalReferenceId(), (i * RECORDS_PER_REQUEST));
+        allMessages.addAll(response.getMessageList().getMessage());
+      }
+    }
+    //Apply all predicates on messages
+    var messages = allMessages.stream()
+            .filter(directionPredicate)
+            .filter(message -> (isEmpty(message.getReceiverList().getReceiver())
+                    || ( message.getReceiverList().getReceiver().stream()
+                    .anyMatch(receiverPredicate))))
+            .collect(Collectors.toList());
+
+    // sort messages
+    sortMessages(messages, messageRequestInfo.getMessageSort(),messageRequestInfo.getMessageSortOrder());
+
+    //convert to message summary
+    return new MessageSummary()
+            .counts(new MessageTotals()
+                    .messagesTotal(messagesResponse.getTotRecords())
+                    .pageTotal(messagesResponse.getReturnedRecords()))
+            .links(getLinks(messages, messageRequestInfo.getPageSize()))
+            .messages(getCatMessages(messages, messageRequestInfo.getMessageRead()
+                    ,jaggaerUserId,messageRequestInfo.getPageSize()));
+  }
+  private Links1 getLinks(final List<uk.gov.crowncommercial.dts.scale.cat.model.jaggaer.Message> messages, final Integer pageSize) {
+    var message = messages.iterator().next();
+    return new Links1().first(URI.create(LINK_URI+(message.getMessageId())))
+            .self(URI.create(LINK_URI+(message.getMessageId())))
+            .prev(URI.create(LINK_URI+(message.getMessageId())))
+            .next(URI.create(LINK_URI+(message.getMessageId()+1)))
+            .last(URI.create(LINK_URI+(message.getMessageId()+pageSize)));
+  }
+
+  private void sortMessages(List<uk.gov.crowncommercial.dts.scale.cat.model.jaggaer.Message> messages, final MessageSort messageSort, MessageSortOrder messageSortOrder) {
+    switch (messageSort) {
+      case TITLE: {
+        if (MessageSortOrder.ASCENDING.equals(messageSortOrder)) {
+          Collections.sort(messages, Comparator.comparing(uk.gov.crowncommercial.dts.scale.cat.model.jaggaer.Message::getSubject));
+        } else {
+          Collections.sort(messages, Comparator.comparing(uk.gov.crowncommercial.dts.scale.cat.model.jaggaer.Message::getSubject).reversed());
+        }
+        break;
+      }
+      case AUTHOR: {
+        if (MessageSortOrder.ASCENDING.equals(messageSortOrder)) {
+          messages.sort(Comparator.comparing(o -> o.getSender().getName()));
+        } else {
+          Collections.sort(messages, comparator.reversed());
+        }
+        break;
+      }
+      default: {
+        if (MessageSortOrder.ASCENDING.equals(messageSortOrder)) {
+          Collections.sort(messages, Comparator.comparing(uk.gov.crowncommercial.dts.scale.cat.model.jaggaer.Message::getSendDate));
+        } else {
+          Collections.sort(messages, Comparator.comparing(uk.gov.crowncommercial.dts.scale.cat.model.jaggaer.Message::getSendDate).reversed());
+        }
+        break;
+      }
+    }
+  }
+
+
+
+  /**
+   * Method to get details of message
+   * @param procId
+   * @param eventId
+   * @param messageId
+   * @param principal
+   * @return Message Summary
+   */
+  public uk.gov.crowncommercial.dts.scale.cat.model.generated.Message getMessageSummary(
+          final Integer procId, final String eventId, final String messageId, final String principal) {
+
+    userProfileService.resolveBuyerUserByEmail(principal)
+            .orElseThrow(() -> new AuthorisationFailureException(JAGGAER_USER_NOT_FOUND)).getUserId();
+    validationService.validateProjectAndEventIds(procId, eventId);
+    var reponse = jaggaerService.getMessage(messageId);
+
+    return new uk.gov.crowncommercial.dts.scale.cat.model.generated.Message()
+            .OCDS(getMessageOCDS(reponse) )
+            .nonOCDS(getMessageNonOCDS(reponse));
+  }
+
+  private List<CaTMessage> getCatMessages(
+          final List<uk.gov.crowncommercial.dts.scale.cat.model.jaggaer.Message> messages,
+          final MessageRead messageRead, final String jaggaerUserId, final Integer pageSize) {
+    return messages.stream().map(message ->  convertMessageToCatMessage(message,messageRead,jaggaerUserId))
+            .limit(pageSize)
+            .collect(Collectors.toList());
+  }
+
+  private CaTMessage convertMessageToCatMessage(
+          final uk.gov.crowncommercial.dts.scale.cat.model.jaggaer.Message message,
+          final MessageRead messageRead, final String jaggaerUserId) {
+
+    final Predicate<Receiver> receiverPredicate = receiver -> MessageRead.READ.equals(messageRead)
+            && receiver.getId().equals(jaggaerUserId);
+    Boolean read;
+    if (CaTMessageNonOCDS.DirectionEnum.SENT.getValue().equals( message.getDirection())) {
+      read = Boolean.FALSE;
+    } else {
+      read = message.getReceiverList().getReceiver().stream()
+              .anyMatch(receiverPredicate);
+    }
+
+    return new CaTMessage().OCDS(getCaTMessageOCDS(message))
+            .nonOCDS(new CaTMessageNonOCDS()
+                    .direction(CaTMessageNonOCDS.DirectionEnum.fromValue(message.getDirection()))
+                    .read(read)
+            );
+  }
+
+  private CaTMessageOCDS getCaTMessageOCDS(uk.gov.crowncommercial.dts.scale.cat.model.jaggaer.Message message) {
+    return new CaTMessageOCDS().date(message.getSendDate()).id(message.getMessageId())
+            .title(message.getSubject())
+            .author(message.getSender().getName());
+  }
+  private MessageOCDS getMessageOCDS(uk.gov.crowncommercial.dts.scale.cat.model.jaggaer.Message message) {
+    return new MessageOCDS().date(message.getSendDate()).id(message.getMessageId())
+            .title(message.getSubject())
+            .author(message.getSender().getName());
+  }
+
+  private MessageNonOCDS getMessageNonOCDS(uk.gov.crowncommercial.dts.scale.cat.model.jaggaer.Message message) {
+    //TODO fix generated class with correct classes
+    return new MessageNonOCDS()
+            .direction(MessageNonOCDS.DirectionEnum.fromValue(message.getDirection()))
+//            .attachments(  message.getAttachmentList().getAttachment().stream()
+//            .map(object -> new MessageNonOCDSAllOfAttachments()).collect(Collectors.toList()))
+            .readList(message.getReadingList().getReading().stream()
+                    .map(reading -> new ContactPoint1().name(reading.getReaderName()))
+                    .collect(Collectors.toList()))
+            .receiverList(message.getReceiverList().getReceiver().stream()
+                    .map(receiver ->  new OrganizationReference1().id(receiver.getId()).name(receiver.getName()))
+                    .collect(Collectors.toList()));
   }
 }
